requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
faiss-cpu==1.8.0